refactor(exchange): move alpaca demo scripts into a demo/ folder
Brings the alpaca broker in line with the trading212 broker, which
already groups its standalone scripts under demo/. `demo.ts` is renamed
to `testAll.ts` to mirror `trading212/demo/testAll.ts` (both files are
comprehensive client-method smoke tests); `fetchCandles.ts` keeps its
name.

This lets the shared root tsconfig.build.json drop two file-by-name
exclude patterns (`**/demo.ts` and `**/fetchCandles.ts`) and rely on
the existing `**/demo` folder pattern alone. Updates the matching
npm scripts and the vitest coverage exclude list.
