+++ packages/exchange/src/broker/alpaca/demo/testAll.ts
@@ -1,7 +1,7 @@
 import 'dotenv-defaults/config';
 import {ms, format} from 'ms';
-import {getAlpacaClient} from './getAlpacaClient.js';
-import {TradingPair} from '../TradingPair.js';
+import {getAlpacaClient} from '../getAlpacaClient.js';
+import {TradingPair} from '../../TradingPair.js';
 
 const exchange = getAlpacaClient({
   apiKey: process.env.ALPACA_PAPER_API_KEY ?? '',
