+++ packages/exchange/vitest.config.ts
@@ -9,12 +9,7 @@ export default mergeConfig(
         // Exclude runnable CLI scripts (top-level await + process.exit) from coverage.
         // They aren't library code and rolldown's coverage parser misreads some TS syntax
         // in them (e.g. type assertions, non-null assertions).
-        exclude: [
-          'fixtures/**',
-          'src/broker/alpaca/demo.ts',
-          'src/broker/alpaca/fetchCandles.ts',
-          'src/broker/trading212/demo/**',
-        ],
+        exclude: ['fixtures/**', 'src/broker/alpaca/demo/**', 'src/broker/trading212/demo/**'],
       },
     },
   })
